Move {en,dis}ableFormElements to features/disable.js
So that all utils have no dependency on others
