+++ src/rails.js
@@ -11,9 +11,9 @@
 import config from './config'
 import { fire, stopEverything } from './utils/event'
 import { refreshCSRFTokens, CSRFProtection, csrfToken, csrfParam } from './utils/csrf'
-import { enableFormElement, enableFormElements, disableFormElement, disableFormElements, blankInputs, nonBlankInputs } from './utils/form'
+import { enableFormElement, disableFormElement, blankInputs, nonBlankInputs } from './utils/form'
 import { href, ajax } from './utils/ajax'
-import { enableElement, disableElement } from './features/disable'
+import { enableElement, disableElement, enableFormElements, disableFormElements } from './features/disable'
 import { allowAction } from './features/confirm'
 import { isRemote, handleRemote } from './features/remote'
 import { handleMethod } from './features/method'
